requirements.txt
allure-pytest==2.15.2
allure-python-commons==2.15.2
attrs==25.4.0
black==25.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
factory_boy==3.3.3
Faker==38.2.0
idna==3.11
iniconfig==2.3.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pycparser==2.23
Pygments==2.19.2
PyMySQL==1.1.2
pytest==9.0.1
pytokens==0.3.0
PyYAML==6.0.3
requests==2.32.5
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
